Break the negative log likelihood into it's own header
