+++ albatross/evaluate.h
@@ -13,70 +13,15 @@
 #ifndef ALBATROSS_EVALUATE_H
 #define ALBATROSS_EVALUATE_H
 
-#include "core/model.h"
-#include "crossvalidation.h"
-#include <Eigen/Cholesky>
-#include <Eigen/Dense>
 #include <functional>
 #include <map>
 #include <math.h>
 #include <memory>
 
+#include "core/dataset.h"
+#include "likelihood.h"
+
 namespace albatross {
-
-/*
- * Negative log likelihood of a univariate normal.
- */
-static inline double negative_log_likelihood(double deviation,
-                                             double variance) {
-  double nll = deviation;
-  nll *= nll;
-  nll /= (2 * variance);
-  nll += 0.5 * log(2 * M_PI * variance);
-  return nll;
-}
-
-static inline double log_sum(const Eigen::VectorXd &x) {
-  double sum = 0.;
-  for (Eigen::Index i = 0; i < x.size(); i++) {
-    sum += log(x[i]);
-  }
-  return sum;
-}
-
-/*
- * Negative log likelihood of a pre decomposed multivariate
- * normal.
- */
-template <typename _MatrixType, int _UpLo>
-static inline double
-negative_log_likelihood(const Eigen::VectorXd &deviation,
-                        const Eigen::LDLT<_MatrixType, _UpLo> &ldlt) {
-  const auto diag = ldlt.vectorD();
-  const double rank = static_cast<double>(diag.size());
-  const double mahalanobis = deviation.dot(ldlt.solve(deviation));
-  const double log_det = log_sum(diag);
-  return 0.5 * (log_det + mahalanobis + rank * log(2 * M_PI));
-}
-
-/*
- * Computes the negative log likelihood under the assumption that the predcitve
- * distribution is multivariate normal.
- */
-static inline double
-negative_log_likelihood(const Eigen::VectorXd &deviation,
-                        const Eigen::MatrixXd &covariance) {
-  assert(deviation.size() == covariance.rows());
-  assert(covariance.cols() == covariance.rows());
-  if (deviation.size() == 1) {
-    // Looks like we have a univariate distribution, skipping
-    // all the matrix decomposition steps should speed this up.
-    return negative_log_likelihood(deviation[0], covariance(0, 0));
-  } else {
-    const auto ldlt = covariance.ldlt();
-    return negative_log_likelihood(deviation, ldlt);
-  }
-}
 
 /*
  * This handles the case where the covariance matrix is diagonal, which
